Rename async-task.ts utils file as types.ts
